Fixed prhmax and derotation
shellscript:
-fixed small bug that limited processing end
-fixed bug that LASTDAY could differ between months

cmorlight:
-enabled processing of prhmax
-fixed error that prevented constant variables to be processed
-fixed error with derotate_uv
-changed deprecated optparse to argparse

+++ src/CMORlight/cmorlight2.py
@@ -19,7 +19,7 @@ from datetime import datetime
 from multiprocessing import Process,Pool
 
 # command line parser
-from optparse import OptionParser
+import argparse
 
 # configuration
 import configuration as config
@@ -55,11 +55,17 @@ def process_resolution(params,reslist):
     var = params[config.get_config_value('index','INDEX_VAR')]
     varRCM=params[config.get_config_value('index','INDEX_RCM_NAME')]
     # create path to input files from basedir,model,driving_model
-    in_dir = "%s/%s" % (tools.get_input_path(var),params[config.get_config_value('index','INDEX_RCM_NAME')])
+    in_dir = "%s/%s" % (tools.get_input_path(),params[config.get_config_value('index','INDEX_RCM_NAME')])
     log.debug("Looking for input dir(1): %s" % (in_dir))
+
+#    #if no input exists for prhmax use input of pr instead
+#    if os.path.isdir(in_dir) == False and var=="prhmax":
+#        varRCM=settings.param["pr"][config.get_config_value('index','INDEX_RCM_NAME')]
+#        log.info("Input directory does not exist. Use variable %s as input " % varRCM)
+#        in_dir = "%s/%s" % (tools.get_input_path(var),varRCM)
     if os.path.isdir(in_dir) == False:
-      log.error("Input directory does not exist(0): %s \n \t Change base path in .ini file or create directory! " % in_dir)
-      return
+        log.error("Input directory does not exist(0): %s \n \t Change base path in .ini file or create directory! " % in_dir)
+        return
 
     multilst=[] #argument list for multiprocessing
     cores=config.get_config_value("integer","cores")
@@ -78,7 +84,7 @@ def process_resolution(params,reslist):
                 log.info("\n###########################################################\n# Var in work: %s / %s\n###########################################################" % (var, varRCM))
                 log.info("Start processing at: "+str(datetime.now()))
             #if limit_range is set: skip file if it is out of range
-            if config.get_config_value('boolean','limit_range') and var not in config.get_model_value('var_list_fixed'):
+            if config.get_config_value('boolean','limit_range') and var not in settings.var_list_fixed:
                 if int(year) < config.get_config_value('integer','proc_start') or int(year) > config.get_config_value('integer','proc_end'):
                     log.debug("File %s out of time range! Skipping ..." % f)
                     continue
@@ -93,8 +99,9 @@ def process_resolution(params,reslist):
                 if os.access(in_file, os.R_OK) == False:
                     log.error("Could not read file '%s', no permission!" % in_file)
                 else:
-                    if var in config.get_model_value('var_list_fixed'):
+                    if var in settings.var_list_fixed:
                         tools.process_file_fix(params,in_file)
+
                     else:
                         if config.get_config_value("boolean","multi"):
                             multilst.append([params,in_file,var,reslist,year])
@@ -108,7 +115,7 @@ def process_resolution(params,reslist):
             i=i+1
 
             #process as many files simultaneously as there are cores specified
-            if (config.get_config_value("boolean","multi") and i==cores):
+            if config.get_config_value("boolean","multi") and i==cores and multilst!=[]:
                 pool=Pool(processes=cores)
                 R=pool.map(process_file_unpack,multilst)
                 pool.terminate()
@@ -133,100 +140,94 @@ def process_resolution(params,reslist):
 def main():
     ''' main program, first read command line parameter '''
 
-    parser = OptionParser(version="%prog "+base.__version__) #VERSION)
+    parser = argparse.ArgumentParser() #VERSION)
 
-    parser.add_option("-i", "--ini",
+    parser.add_argument("-i", "--ini",
                             action="store", dest = "inifile", default = "control_cmor2.ini",
                             help = "configuration file (.ini)")
-    parser.add_option("-p", "--param",
+    parser.add_argument("-p", "--param",
                             action="store", dest = "paramfile", default = "",
                             help = "model parameter file (table)")
-    parser.add_option("-r", "--resolution",
+    parser.add_argument("-r", "--resolution",
                             action="store", dest = "reslist", default = "",
                             help = "list of desired output resolutions (supported: 1hr (1-hourly), 3hr (3-hourly),6hr (6-hourly),day (daily),mon (monthly) ,sem (seasonal),fx (for time invariant variables)")
-    parser.add_option("-v", "--varlist",
+    parser.add_argument("-v", "--varlist",
                             action="store", dest = "varlist", default = "pr",
                             help = "list of variables to be processed")
-    parser.add_option("-a", "--all",
+    parser.add_argument("-a", "--all",
                             action="store_true", dest = "all_vars", default = False,
                             help = "process all variables")
-    parser.add_option("-c", "--chunk-var",
+    parser.add_argument("-c", "--chunk-var",
                             action="store_true", dest="chunk_var", default = False,
                             help="go call chunking for the variable list")
-#    parser.add_option("-s", "--seasonal-mean",
+#    parser.add_argument("-s", "--seasonal-mean",
 #                            action="store_true", dest="seasonal_mean", default=False,
 #                            help="go calculate seasonal mean from aggregated monthly data")
-    parser.add_option("-n", "--use-version",
+    parser.add_argument("-n", "--use-version",
                             action="store", dest = "use_version", default = tools.new_dataset_version(),
                             help = "version for drs (default: today in format YYYYMMDD)")
-    parser.add_option("-d", "--derotate-uv",
-                            action="store_true", dest = "derotate_uv", default=False,
+    parser.add_argument("-d", "--no_derotate",
+                            action="store_false", dest = "derotate_uv", default=True,
                             help = "derotate all u and v avariables")
-    parser.add_option("-t", "--test-var",
+    parser.add_argument("-t", "--test-var",
                             action="store_true", dest="test_var", default = False,
                             help="test possible resolution for all vars")
-    parser.add_option("-k", "--corr-var",
+    parser.add_argument("-k", "--corr-var",
                             action="store_true", dest="corr_var", default = False,
                             help="correct variable by corr_key")
-    parser.add_option("-o", "--corr-key",
+    parser.add_argument("-o", "--corr-key",
                             action="store", dest="corr_key", default = "climatology",
                             help="correct key to select a function")
-    parser.add_option("-y", "--alt-start-year",
+    parser.add_argument("-y", "--alt-start-year",
                             action="store", dest="alt_start_year", default = 2100,
                             help="use alternate start year")
-    parser.add_option("-u", "--use-alt-units",
+    parser.add_argument("-u", "--use-alt-units",
                             action="store_true", dest="use_alt_units", default = False,
                             help="use alternate units for input data (only day and mon)")
-    parser.add_option("-m", "--model",
+    parser.add_argument("-m", "--model",
                            action="store", dest="act_model", default = 'CCLM',
                           help="set used model (supported: [default: CCLM],WRF)")
-    parser.add_option("-g", "--gcm_driving_model",
+    parser.add_argument("-g", "--gcm_driving_model",
                             action="store", dest="driving_model_id", default = "",
                             help="set used driving model")
-    parser.add_option("-x", "--experiment",
+    parser.add_argument("-x", "--experiment",
                             action="store", dest="driving_experiment_name", default = "",
                             help="set used experiment")
-    parser.add_option("-E", "--ensemble",
+    parser.add_argument("-E", "--ensemble",
                             action="store", dest="driving_model_ensemble_member", default = "",
                             help="set used ensemble")
-    parser.add_option("-O", "--overwrite",
+    parser.add_argument("-O", "--overwrite",
                             action="store_true", dest="overwrite", default = False,
                             help="Overwrite existent output files")
-    parser.add_option("-f", "--force_proc",
+    parser.add_argument("-f", "--force_proc",
                             action="store_false", dest="process_table_only", default = True,
                             help="Try to process variable at specific resolution regardless of what is written in the parameter table")
-    parser.add_option("-S", "--silent",
+    parser.add_argument("-S", "--silent",
                             action="store_false", dest="normal_log", default = True,
                             help="Write only minimal information to log (variables and resolutions in progress, warnings and errors)")
-    parser.add_option("-V", "--verbose",
+    parser.add_argument("-V", "--verbose",
                             action="store_true", dest="verbose_log", default = False,
                             help="Verbose logging for debugging")
-    parser.add_option("-A", "--append_log",
+    parser.add_argument("-A", "--append_log",
                             action="store_true", dest="append_log", default = False,
                             help="Append to log instead of overwrite")
-    parser.add_option("-l", "--limit",
+    parser.add_argument("-l", "--limit",
                             action="store_true", dest="limit_range", default = False,
                             help="Limit time range for processing (range set in .ini file or parsed)")
-    parser.add_option("-s", "--start",
+    parser.add_argument("-s", "--start",
                             action="store", dest="proc_start", default = "",
                             help="Start year for processing if --limit is set.")
-    parser.add_option("-e", "--end",
+    parser.add_argument("-e", "--end",
                             action="store", dest="proc_end", default = "",
                             help="End year for processing if --limit is set.")
-    parser.add_option("-M", "--multi",
+    parser.add_argument("-M", "--multi",
                             action="store_true", dest="multi", default = False,
                             help="Use multiprocessing with number of cores specified in .ini file.")
-    parser.add_option( "--remove",
+    parser.add_argument( "--remove",
                             action="store_true", dest="remove_src", default = False,
                             help="Remove source files after chunking")
 
-
-
-
-
-
-
-    (options, args) = parser.parse_args()
+    options = parser.parse_args()
     options_dict=vars(options)
     config.load_configuration(options.inifile)
 
@@ -239,18 +240,26 @@ def main():
         options.limit_range=True
 
    #store parsed arguments in config
-    setval=["paramfile","driving_model_id","driving_experiment_name","driving_model_ensemble_member"]
+    setval_settings_model=["paramfile","driving_model_id","driving_experiment_name","driving_model_ensemble_member"]
+    setval_integer=["proc_start","proc_end"]
+    setval=[]
+    setval.extend(setval_settings_model)
+    setval.extend(setval_integer)
     for val in setval:
         if options_dict[val]!="":
-            config.set_config_value('settings_',val,options_dict[val])
+            if val in setval_settings_model:
+                config.set_config_value('settings_',val,options_dict[val])
+            elif val in setval_integer:
+                config.set_config_value('integer',val,options_dict[val])
+
 
     config.set_config_value('settings','model',options.act_model)
     config.set_config_value('boolean','overwrite',options.overwrite)
     config.set_config_value('boolean','limit_range',options.limit_range)
     config.set_config_value('boolean','remove_src',options.remove_src)
     config.set_config_value('boolean','multi',options.multi)
-    config.set_config_value('integer','proc_start',options.proc_start)
-    config.set_config_value('integer','proc_end',options.proc_end)
+    config.set_config_value('boolean','derotate_uv',options.derotate_uv)
+
 
     process_list=[config.get_model_value('driving_model_id'),config.get_model_value('driving_experiment_name'),config.get_model_value('driving_model_ensemble_member')]
     # now read paramfile for all variables for this RCM
@@ -297,10 +306,7 @@ def main():
     if config.get_config_value('boolean','add_version_to_outpath'):
         settings.use_version = "v%s" % (options.use_version)
     settings.use_alt_units = options.use_alt_units
-    # derotate u and v
-    if options.derotate_uv == True:
-        tools.derotate_uv()
-        return
+
 
     if options.all_vars == False:
         varlist = options.varlist.split(',')
@@ -352,8 +358,8 @@ def main():
             continue
         log.log(35,"\n\n\n###########################################################\n# Var in work: %s / %s\n###########################################################" % (varCMOR, varRCM))
 
-        if var not in config.get_model_value('var_list_fixed'):
-            reslist_act=list(reslist) #new copy of reslist
+        reslist_act=list(reslist) #new copy of reslist
+        if (varRCM not in settings.var_list_fixed) and (varCMOR not in settings.var_list_fixed):
             for res in reslist:
                 if tools.check_resolution(params,res,options.process_table_only) == False:
                     reslist_act.remove(res) #remove resolution from list (for this variable) if it is not in table or if it is not supported
@@ -365,9 +371,8 @@ def main():
             continue
         # set global attributes in the dictionary
         tools.set_attributes(params,process_list)
-
         # skip fixed fields from chunking, makes no sense to chunk
-        if options.chunk_var == True and not var in config.get_model_value('var_list_fixed'):
+        if options.chunk_var == True and not var in settings.var_list_fixed:
             log.log(35, "Chunking files \n #######################")
             tools.proc_chunking(params,reslist_act)
 
@@ -387,10 +392,12 @@ def main():
 if __name__ == "__main__":
     ''' main program '''
 
+    #start timing
+    time1=datetime.now()
     # call main function
     main()
     #clean up temp files
-    shutil.rmtree(settings.DirWork)
+   # shutil.rmtree(settings.DirWork)
 
   #  Clean up .tmp files in output directory (if on last run program crashed while writing a file)
     for root, dirs, files in os.walk(settings.DirOut):
@@ -398,8 +405,15 @@ if __name__ == "__main__":
             if file[-4:]==".tmp":
                 os.remove(os.path.join(root, file))
 
+    time2=datetime.now()
     log = logging.getLogger('cmorlight')
     log.propagate = True
+    #calculate processing time
+    time_diff=time2-time1
+    hours, remainder = divmod(time_diff.total_seconds(), 3600)
+    minutes, seconds = divmod(remainder, 60)
+
+    log.log(35,'Total processing time: %s hours %s minutes %s seconds' % (int(hours),int(minutes),round(seconds)))
     log.log(35,'\n##################################\n########  End of script.  ########\n##################################')
     ######################
     # END of program!!!  #
